requirements.txt
anthropic==0.37.1
datasets==2.21.0
openai==1.52.2
protobuf==5.28.3
PyYAML==6.0.2
sentence_transformers==3.1.1
torch==2.5.1
google-generativeai==0.8.3

safetensors==0.5.2
huggingface-hub==0.27.1
transformers==4.48.1
numpy==2.2.2

